requirements.txt
# Core dependencies
numpy==1.26.0
opencv-python==4.10.0.84
Pillow==10.1.0
ultralytics==8.3.99

# Web scraping
selenium==4.15.2

# Database
psycopg2-binary==2.9.10

# Configuration
python-dotenv==1.0.0
